+++ internal/database/sqliterepository.go
@@ -44,6 +44,7 @@ var keys = []keyDefinition{
 	{"release", "TEXT NOT NULL"},
 	{"prerelease", "INTEGER NOT NULL"},
 	{"kibanaVersion", "TEXT NOT NULL"},
+	{"capabilities", "TEXT NOT NULL"},
 	{"type", "TEXT NOT NULL"},
 	{"path", "TEXT NOT NULL"},
 	{dataColumnName, "BLOB NOT NULL"},
@@ -226,6 +227,7 @@ func (r *SQLiteRepository) BulkAdd(ctx context.Context, database string, pkgs []
 				pkgs[i].Release,
 				pkgs[i].Prerelease,
 				pkgs[i].KibanaVersion,
+				pkgs[i].Capabilities,
 				pkgs[i].Type,
 				pkgs[i].Path,
 				pkgs[i].Data,
@@ -291,6 +293,8 @@ func (r *SQLiteRepository) AllFunc(ctx context.Context, database string, whereOp
 		case k.Name == dataColumnName && useBaseData:
 			continue
 		case k.Name == baseDataColumnName && !useBaseData:
+			continue
+		case k.Name == "capabilities":
 			continue
 		default:
 			getKeys = append(getKeys, k.Name)
@@ -372,6 +376,7 @@ type FilterOptions struct {
 	KibanaVersion string
 	SpecMin       string
 	SpecMax       string
+	Capabilities  []string
 	// It cannot be filtered by capabilities at database level, since it would be
 	// complicated using SQL logic to ensure that all the capabilities defined in the package
 	// are present in the query filter.
@@ -463,6 +468,16 @@ func (o *SQLOptions) Where() (string, []any) {
 		}
 		sb.WriteString("semver_compare_le(formatVersionMajorMinor, ?) = 1")
 		args = append(args, o.Filter.SpecMax)
+	}
+
+	if len(o.Filter.Capabilities) > 0 {
+		if sb.Len() > 0 {
+			sb.WriteString(" AND ")
+		}
+
+		queryCapabilities := strings.Join(o.Filter.Capabilities, ",")
+		sb.WriteString("( capabilities == '' OR all_capabilities_are_supported(capabilities, ?) = 1 )")
+		args = append(args, queryCapabilities)
 	}
 
 	if sb.String() == "" {