+++ internal/storage/sqlindexer_test.go
@@ -172,6 +172,16 @@ func BenchmarkSQLIndexerGet(b *testing.B) {
 			Prerelease:    false,
 			KibanaVersion: semver.MustParse("9.0.0"),
 		}})
+		indexer.Get(b.Context(), &packages.GetOptions{Filter: &packages.Filter{
+			AllVersions:  false,
+			Prerelease:   false,
+			Capabilities: []string{"security", "observability"},
+		}})
+		indexer.Get(b.Context(), &packages.GetOptions{Filter: &packages.Filter{
+			AllVersions:  false,
+			Prerelease:   false,
+			Capabilities: []string{"apm"},
+		}})
 	}
 }